+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/assembly/KafkaAssemblyOperatorTest.java
@@ -146,10 +146,6 @@ public class KafkaAssemblyOperatorTest {
             .withNewMetadata()
                 .withName(CLUSTER_NAME)
                 .withNamespace(NAMESPACE)
-                .withAnnotations(Map.of(
-                        Annotations.ANNO_STRIMZI_IO_NODE_POOLS, "enabled",
-                        Annotations.ANNO_STRIMZI_IO_KRAFT, "enabled"
-                ))
                 .withGeneration(1L)
             .endMetadata()
             .withNewSpec()